move fitToClass function from page.vue into store. Fix pdf fitTo Orignal display to be like Width, fix fullpdf fitTo Display.
