+++ frontend/src/components/reader/pages/page/page.vue
@@ -10,7 +10,6 @@
       v-else
       ref="pageComponent"
       :book="book"
-      :fit-to-class="fitToClass"
       :page="1"
       :src="src"
       @error="onError"
@@ -22,7 +21,6 @@
 
 <script>
 import { mapActions, mapGetters, mapState } from "pinia";
-import titleize from "titleize";
 import { defineAsyncComponent, markRaw } from "vue";
 
 import { getComicPageSource } from "@/api/v3/reader";
@@ -35,8 +33,6 @@ import ErrorPage from "@/components/reader/pages/page/page-error.vue";
 import ImgPage from "@/components/reader/pages/page/page-img.vue";
 
 const PROGRESSS_DELAY_MS = 333;
-
-const FIT_TO_CHOICES = { S: "Screen", W: "Width", H: "Height", O: "Original" };
 
 export default {
   name: "BookPage",
@@ -76,24 +72,6 @@ export default {
       s.height = img.naturalHeight * this.scale + "px";
       s.width = img.naturalWidth * this.scale + "px";
       return s;
-    },
-    fitToClass() {
-      const classes = {};
-      if (this.scale > 1) {
-        return classes;
-      }
-      const fitTo = FIT_TO_CHOICES[this.settings.fitTo];
-      if (fitTo) {
-        let fitToClass = "fitTo";
-        fitToClass += titleize(fitTo);
-        if (this.isVertical) {
-          fitToClass += "Vertical";
-        } else if (this.settings.twoPages) {
-          fitToClass += "Two";
-        }
-        classes[fitToClass] = true;
-      }
-      return classes;
     },
     src() {
       const params = {
